AgencyService
*

Creating and implemting service for agency

*

+++ src/main/java/br/com/vsgdev/emotionsRoute/model/User.java
@@ -22,12 +22,12 @@ import com.fasterxml.jackson.annotation.JsonProperty.Access;
 import br.com.vsgdev.emotionsRoute.enums.UserTypeEnum;
 
 @Entity
-@Inheritance( strategy = InheritanceType.TABLE_PER_CLASS )
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 public abstract class User {
 
 	@Id
 	@GeneratedValue
-	private Long id;
+	private int id;
 
 	@NotNull
 	private String name;
@@ -37,29 +37,29 @@ public abstract class User {
 	private List<String> phones;
 
 	@NotNull
-	@Column( unique = true )
+	@Column(unique = true)
 	private String email;
 
 	@NotNull
-	@OneToOne( fetch = FetchType.EAGER, cascade = CascadeType.ALL )
+	@OneToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
 	private Address address;
 
-	@JsonProperty( access = Access.WRITE_ONLY )
+	@JsonProperty(access = Access.WRITE_ONLY)
 	private String password;
 
 	@NotNull
-	@Enumerated( EnumType.STRING )
+	@Enumerated(EnumType.STRING)
 	private UserTypeEnum userType;
 
 	private double rate;
 
 	private boolean active = true;
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId( Long id ) {
+	public void setId(int id) {
 		this.id = id;
 	}
 
@@ -67,7 +67,7 @@ public abstract class User {
 		return name;
 	}
 
-	public void setName( String name ) {
+	public void setName(String name) {
 		this.name = name;
 	}
 
@@ -75,7 +75,7 @@ public abstract class User {
 		return phones;
 	}
 
-	public void setPhones( List<String> phones ) {
+	public void setPhones(List<String> phones) {
 		this.phones = phones;
 	}
 
@@ -83,7 +83,7 @@ public abstract class User {
 		return email;
 	}
 
-	public void setEmail( String email ) {
+	public void setEmail(String email) {
 		this.email = email;
 	}
 
@@ -91,7 +91,7 @@ public abstract class User {
 		return address;
 	}
 
-	public void setAddress( Address address ) {
+	public void setAddress(Address address) {
 		this.address = address;
 	}
 
@@ -99,7 +99,7 @@ public abstract class User {
 		return password;
 	}
 
-	public void setPassword( String password ) {
+	public void setPassword(String password) {
 		this.password = password;
 	}
 
@@ -107,7 +107,7 @@ public abstract class User {
 		return userType;
 	}
 
-	public void setUserType( UserTypeEnum userType ) {
+	public void setUserType(UserTypeEnum userType) {
 		this.userType = userType;
 	}
 
@@ -115,7 +115,7 @@ public abstract class User {
 		return rate;
 	}
 
-	public void setRate( double rate ) {
+	public void setRate(double rate) {
 		this.rate = rate;
 	}
 
@@ -123,7 +123,7 @@ public abstract class User {
 		return active;
 	}
 
-	public void setActive( boolean active ) {
+	public void setActive(boolean active) {
 		this.active = active;
 	}
